+++ backend/users/src/app.controller.spec.ts
@@ -1,22 +1,22 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { Test, TestingModule } from "@nestjs/testing";
+import { UserController } from "./user.controller";
+import { AppService } from "./user.service";
 
-describe('AppController', () => {
-  let appController: AppController;
+describe("UserController", () => {
+    let appController: UserController;
 
-  beforeEach(async () => {
-    const app: TestingModule = await Test.createTestingModule({
-      controllers: [AppController],
-      providers: [AppService],
-    }).compile();
+    beforeEach(async () => {
+        const app: TestingModule = await Test.createTestingModule({
+            controllers: [UserController],
+            providers: [AppService],
+        }).compile();
 
-    appController = app.get<AppController>(AppController);
-  });
-
-  describe('root', () => {
-    it('should return "Hello World!"', () => {
-      expect(appController.getHello()).toBe('Hello World!');
+        appController = app.get<UserController>(UserController);
     });
-  });
+
+    describe("root", () => {
+        it('should return "Hello World!"', () => {
+            expect(appController.getHello()).toBe("Hello World!");
+        });
+    });
 });
