+++ backend/users/src/user.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from "@nestjs/common";
+
+@Injectable()
+export class UserService {
+    getHello(): string {
+        return "Hello World!";
+    }
+}
